backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-limiter>=4.0
bcrypt==4.1.3
PyJWT==2.8.0
openpyxl==3.1.3
requests==2.32.3
cryptography==42.0.8
wechatpayv3==1.2.31
